backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudevents==1.12.0
colorama==0.4.6
cryptography==46.0.2
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.118.0
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib-metadata==8.7.0
isodate==0.7.2
jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
markupsafe==3.0.3
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.3.3
openai==2.0.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
parse==1.20.2
pathable==0.4.4
prance==25.4.8.0
propcache==0.3.2
protobuf==6.32.1
pybars4==0.9.13
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pyee==13.0.0
pyjwt==2.10.1
pylibsrtp==0.12.0
pymeta3==0.5.1
pymongo==4.15.2
pyopenssl==25.3.0
python-dotenv==1.1.1
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rpds-py==0.27.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
scipy==1.16.2
semantic-kernel==1.36.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
